bring phpunit test up to psr4 standard
