+++ tests/phpunit/va_gov_form_builder/functional/pages/ViewFormTest.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace tests\phpunit\va_gov_form_builder\functional\content_pages;
+namespace Tests\va_gov_form_builder\functional\pages;
 
 use tests\phpunit\va_gov_form_builder\Traits\TestPageLoads;
 use Tests\Support\Classes\VaGovExistingSiteBase;
 
 /**
- * Functional test of the Review-and-sign page.
+ * Functional test of the View-form page.
  *
  * @group functional
  * @group all
  */
-class ReviewAndSignTest extends VaGovExistingSiteBase {
+class ViewFormTest extends VaGovExistingSiteBase {
   use TestPageLoads;
 
   /**
@@ -30,7 +30,7 @@ class ReviewAndSignTest extends VaGovExistingSiteBase {
    * Returns the url for this page.
    */
   private function getPageUrl() {
-    return "/form-builder/{$this->digitalFormNode->id()}/review-and-sign";
+    return "/form-builder/{$this->digitalFormNode->id()}/view-form";
   }
 
   /**
@@ -73,7 +73,7 @@ class ReviewAndSignTest extends VaGovExistingSiteBase {
    */
   public function testPageLoads() {
     // Ensure page loads.
-    $this->sharedTestPageLoads($this->getPageUrl(), 'Reviewing the submission');
+    $this->sharedTestPageLoads($this->getPageUrl(), 'Reviewing the form');
   }
 
   /**
@@ -95,36 +95,7 @@ class ReviewAndSignTest extends VaGovExistingSiteBase {
     );
   }
 
-  /**
-   * Test that the page has the expected breadcrumbs.
-   */
-  public function testPageBreadcrumbs() {
-    $this->sharedTestPageHasExpectedBreadcrumbs(
-      $this->getPageUrl(),
-      [
-        [
-          'label' => 'Home',
-          'url' => '/form-builder/home',
-        ],
-        [
-          'label' => $this->digitalFormNode->getTitle(),
-          'url' => "/form-builder/{$this->digitalFormNode->id()}",
-        ],
-        [
-          'label' => 'Review page',
-          'url' => "#content",
-        ],
-      ],
-    );
-  }
-
-  /**
-   * Test the primary button.
-   */
-  public function testPrimaryButton() {
-    $this->drupalGet($this->getPageUrl());
-    $this->click('a#form-builder-primary-button');
-    $this->assertSession()->addressEquals("/form-builder/{$this->digitalFormNode->id()}");
-  }
-
+  // Cannot test these until we have the staging-url field.
+  // @todo test breadcrumbs.
+  // @todo test buttons.
 }
